requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
braceexpand==0.1.7
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
DataProperty==1.0.1
datasets==2.16.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.2
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.14.0
flash-attn==2.5.9.post1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.63.0
gspread==6.1.2
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.0
huggingface-hub==0.23.4
idna==3.4
importlib_resources==6.4.0
ipykernel==6.29.4
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
linkify-it-py==2.0.3
lit==15.0.7
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numexpr==2.10.0
numpy==1.26.3
nvidia-nccl-cu12==2.18.1
oauth2client==4.1.3
oauthlib==3.2.2
# Editable install with no version control (open_clip_torch==2.29.0)

openai==0.27.8
openpyxl==3.1.3
orjson==3.9.15
packaging==24.0
pandas==2.2.1
parso==0.8.4
pathvalidate==3.2.0
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==8.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
PySocks==1.7.1
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==2.0.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
rsa==4.9
ruff==0.4.2
sacrebleu==2.4.1
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
timm==1.0.12
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.4
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.33.0
triton==2.1.0
trl==0.8.6
typepy==1.3.2
typer==0.12.3
typing_extensions==4.8.0
tyro==0.8.3
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.29.0
wandb==0.16.5
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.2
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0